perf(cli): load optional features on demand
The root entry statically reached completion generation, definition and
input schema generation, config and manifest discovery, and the terminal
prompter, so every program paid for all of them at import time. Those
now live on their own subpath entries, `@kjanat/dreamcli/completion`,
`@kjanat/dreamcli/json-schema`, `@kjanat/dreamcli/config`, and
`@kjanat/dreamcli/prompt`, and the builder reaches them with dynamic
imports: `.completions()` and `--completions` load the generators,
`--help --json` loads the schema module, `.run()` loads discovery when
`.config()` or `.manifest()` is active and the terminal prompter when
stdin is a TTY. Every type stays on the root entry, as do `SHELLS`, the
schema constants, and `resolvePromptConfig()`.

The build groups every hot-path module into one `core` chunk, so a plain
command loads two files. Under Node a plain command loads `index.mjs`
and `core`, a completions call adds the completion chunk, and JSON help
adds the schema chunk. The root entry's static closure drops from 247 kB
in 4 files to 179 kB in 2, and a Node import of it from 14 ms to 11 ms.

Part of #130.
