+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   Stdin, environment, config, prompt, default, and argv failures now follow the same schema-controlled policy instead of treating non-argv sources as secret. Messages and structured details retain non-sensitive values, collection keys, validator paths, filesystem paths, and adapter causes. Mark every secret-bearing input `.sensitive()`; its raw-derived fields are omitted while source, type, constraint, and allowed-value metadata remains. `CLIError.toJSON()` projects `details` onto JSON-representable values so `--json` reporting cannot throw on a retained runtime value: bigints serialize as decimal digits and entries JSON cannot carry, such as cyclic references, are omitted. The error object keeps the runtime value.
 - **Smaller npm package, faster startup** (https://github.com/kjanat/dreamcli/issues/130).\
   The unpacked package shrinks from 1.5 MB to about 0.7 MB. Runtime JavaScript is bundled into the four public entrypoints plus shared chunks and minified with module-level names kept, so stack traces still name framework functions. Emitted JavaScript no longer carries a copy of the API documentation; the declaration files keep it. Declarations are generated for the public entrypoints only. The tarball no longer ships `examples/` or `CHANGELOG.md`; both remain in the repository and the docs site. `package.json` declares `sideEffects: false` so bundlers drop unused framework modules. `runner run package:check` enforces size and hot-path budgets and runs as part of `runner run bd`.
+- **Breaking: optional features moved to subpath entries and load on demand** (https://github.com/kjanat/dreamcli/issues/130).\
+  The root entry no longer exports the completion generators, `generateSchema()`, `generateCommandSchema()`, `generateInputSchema()`, `definitionMetaSchema`, config and manifest discovery, or `createTerminalPrompter()`. Import them from `@kjanat/dreamcli/completion`, `@kjanat/dreamcli/json-schema`, `@kjanat/dreamcli/config`, and `@kjanat/dreamcli/prompt`. Every type stays on the root entry, as do `SHELLS`, `DEFINITION_SCHEMA_URL`, `DEFINITION_SCHEMA_VERSION`, and `resolvePromptConfig()`. `cli()` loads each feature with a dynamic import when a run needs it, so importing the root entry pulls in about a quarter less code and a program that never renders a completion script, serializes `--help --json`, discovers config, or prompts never loads that code. See the v4 upgrade guide for the import table.
 
 ### Release checklist
