+++ docs/.vitepress/config.ts
@@ -252,6 +252,10 @@ export default defineConfig({
 						{ text: '@kjanat/dreamcli', link: '/reference/main' },
 						{ text: '@kjanat/dreamcli/testkit', link: '/reference/testkit' },
 						{ text: '@kjanat/dreamcli/runtime', link: '/reference/runtime' },
+						{ text: '@kjanat/dreamcli/completion', link: '/reference/completion' },
+						{ text: '@kjanat/dreamcli/json-schema', link: '/reference/json-schema' },
+						{ text: '@kjanat/dreamcli/config', link: '/reference/config' },
+						{ text: '@kjanat/dreamcli/prompt', link: '/reference/prompt' },
 					],
 				},
 			],
