+++ docs/.vitepress/data/typedoc.test.ts
@@ -30,6 +30,38 @@ describe('typedoc normalization', () => {
 				missingExports: [],
 			},
 			{
+				entrypoint: '@kjanat/dreamcli/completion',
+				subpath: './completion',
+				sourcePath: 'src/completion.ts',
+				hasTypeDoc: true,
+				exportIds: expect.arrayContaining(['@kjanat/dreamcli/completion:generateCompletion']),
+				missingExports: [],
+			},
+			{
+				entrypoint: '@kjanat/dreamcli/config',
+				subpath: './config',
+				sourcePath: 'src/config.ts',
+				hasTypeDoc: true,
+				exportIds: expect.arrayContaining(['@kjanat/dreamcli/config:discoverConfig']),
+				missingExports: [],
+			},
+			{
+				entrypoint: '@kjanat/dreamcli/json-schema',
+				subpath: './json-schema',
+				sourcePath: 'src/json-schema.ts',
+				hasTypeDoc: true,
+				exportIds: expect.arrayContaining(['@kjanat/dreamcli/json-schema:generateSchema']),
+				missingExports: [],
+			},
+			{
+				entrypoint: '@kjanat/dreamcli/prompt',
+				subpath: './prompt',
+				sourcePath: 'src/prompt.ts',
+				hasTypeDoc: true,
+				exportIds: expect.arrayContaining(['@kjanat/dreamcli/prompt:createTerminalPrompter']),
+				missingExports: [],
+			},
+			{
 				entrypoint: '@kjanat/dreamcli/runtime',
 				subpath: './runtime',
 				sourcePath: 'src/runtime.ts',
