+++ package.json
@@ -29,6 +29,10 @@
 	"type": "module",
 	"imports": {
 		"#dreamcli": "./src/index.ts",
+		"#dreamcli/completion": "./src/completion.ts",
+		"#dreamcli/config": "./src/config.ts",
+		"#dreamcli/json-schema": "./src/json-schema.ts",
+		"#dreamcli/prompt": "./src/prompt.ts",
 		"#dreamcli/runtime": "./src/runtime.ts",
 		"#dreamcli/testkit": "./src/testkit.ts",
 		"#internals/*": "./src/*",
@@ -39,6 +43,22 @@
 		".": {
 			"types": "./dist/index.d.mts",
 			"default": "./dist/index.mjs"
+		},
+		"./completion": {
+			"types": "./dist/completion.d.mts",
+			"default": "./dist/completion.mjs"
+		},
+		"./config": {
+			"types": "./dist/config.d.mts",
+			"default": "./dist/config.mjs"
+		},
+		"./json-schema": {
+			"types": "./dist/json-schema.d.mts",
+			"default": "./dist/json-schema.mjs"
+		},
+		"./prompt": {
+			"types": "./dist/prompt.d.mts",
+			"default": "./dist/prompt.mjs"
 		},
 		"./runtime": {
 			"types": "./dist/runtime.d.mts",
